refactor(standup): share response validation contracts
